separate timer logic from move
also multiple fixes, refactors, more compile warnings

+++ include/sdl_utils.h
@@ -3,7 +3,7 @@
 #include <SDL_render.h>
 #include <cstdint>
 
-constexpr float PI = 3.14159265358979323846;
+constexpr float PI = 3.14159265358979323846f;
 
 void SDL_DrawCircle(SDL_Renderer* renderer, int32_t cx, int32_t cy,
                     int32_t radius);
@@ -12,10 +12,10 @@ void SDL_DrawDisk(SDL_Renderer* renderer, int32_t cx, int32_t cy,
                   int32_t radius);
 
 struct Color {
-  int r;
-  int g;
-  int b;
-  int a;
+  uint8_t r;
+  uint8_t g;
+  uint8_t b;
+  uint8_t a;
 
   Color operator+(const Color& other) const;
   Color operator-(const Color& other) const;
